Move up RemoveBreakpoint command
